# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/nodes/{id}": {
      "get": {
        "operationId": "NodesController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeDetailDto"
                }
              }
            }
          }
        },
        "summary": "Node metadata with breadcrumbs and what the caller may do",
        "tags": [
          "nodes"
        ]
      },
      "patch": {
        "operationId": "NodesController_rename",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameNodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeDto"
                }
              }
            }
          }
        },
        "summary": "Rename. Renaming a data room’s root renames the room.",
        "tags": [
          "nodes"
        ]
      },
      "delete": {
        "description": "Storage objects are queued for removal in the same transaction and deleted afterwards.",
        "operationId": "NodesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete a node and everything under it",
        "tags": [
          "nodes"
        ]
      }
    },
    "/nodes/{id}/children": {
      "get": {
        "description": "Keyset pagination: pass the previous page’s cursor. Ordered folders first, then by name, case-insensitively.",
        "operationId": "NodesController_children",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor from the previous page",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "FOLDER",
                "FILE"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChildrenPageDto"
                }
              }
            }
          }
        },
        "summary": "One page of a folder’s contents",
        "tags": [
          "nodes"
        ]
      }
    },
    "/nodes/{id}/stats": {
      "get": {
        "description": "Used by the folder header and by the delete confirmation, so the warning states what will actually be removed.",
        "operationId": "NodesController_stats",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubtreeStatsDto"
                }
              }
            }
          }
        },
        "summary": "Totals for everything under this node",
        "tags": [
          "nodes"
        ]
      }
    },
    "/nodes/{id}/search": {
      "get": {
        "description": "Scoped to the node it is called on, so a guest searching a shared folder cannot reach anything outside it — the same endpoint serves both without a separate rule.",
        "operationId": "NodesController_search",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": true,
            "in": "query",
            "description": "At least three characters: shorter queries produce no complete trigrams, so the index cannot serve them and every keystroke would scan the room.",
            "schema": {
              "minLength": 3,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchHitDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Find items by name inside this subtree",
        "tags": [
          "nodes"
        ]
      }
    },
    "/nodes/folders": {
      "post": {
        "operationId": "NodesController_createFolder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFolderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeDto"
                }
              }
            }
          }
        },
        "tags": [
          "nodes"
        ]
      }
    },
    "/nodes/{id}/move": {
      "post": {
        "operationId": "NodesController_move",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveNodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeDto"
                }
              }
            }
          }
        },
        "tags": [
          "nodes"
        ]
      }
    },
...
  },
  "info": {
    "title": "Data Room API",
    "description": "Secure document repository for due diligence.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "NodeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "dataRoomId": {
            "type": "string"
          },
          "parentId": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "enum": [
              "FOLDER",
              "FILE"
            ]
          },
          "name": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "sizeBytes": {
            "type": "number",
            "nullable": true,
            "description": "Current version size"
          },
          "mimeType": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "dataRoomId",
          "parentId",
          "type",
          "name",
          "updatedAt"
        ]
      },
      "BreadcrumbDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "CapabilitiesDto": {
        "type": "object",
        "properties": {
          "canCreate": {
            "type": "boolean"
          },
          "canRename": {
            "type": "boolean"
          },
          "canMove": {
            "type": "boolean"
          },
          "canDelete": {
            "type": "boolean"
          },
          "canUpload": {
            "type": "boolean"
          },
          "canShare": {
            "type": "boolean"
          },
          "canViewVersionHistory": {
            "type": "boolean"
          }
        },
        "required": [
          "canCreate",
          "canRename",
          "canMove",
          "canDelete",
          "canUpload",
          "canShare",
          "canViewVersionHistory"
        ]
      },
      "NodeDetailDto": {
        "type": "object",
        "properties": {
          "node": {
            "$ref": "#/components/schemas/NodeDto"
          },
          "breadcrumbs": {
            "description": "Root first; truncated for guests",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BreadcrumbDto"
            }
          },
          "capabilities": {
            "$ref": "#/components/schemas/CapabilitiesDto"
          }
        },
        "required": [
          "node",
          "breadcrumbs",
          "capabilities"
        ]
      },
      "ChildrenPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NodeDto"
            }
          },
          "nextCursor": {
            "type": "string",
            "description": "Absent when this is the last page"
          }
        },
        "required": [
          "items"
        ]
      },
      "SubtreeStatsDto": {
        "type": "object",
        "properties": {
          "folderCount": {
            "type": "number"
          },
          "fileCount": {
            "type": "number"
          },
          "totalSizeBytes": {
            "type": "number",
            "description": "Sum of current versions — the logical size"
          }
        },
        "required": [
          "folderCount",
          "fileCount",
          "totalSizeBytes"
        ]
      },
      "SearchHitDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "dataRoomId": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "FOLDER",
              "FILE"
            ]
          },
          "name": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "sizeBytes": {
            "type": "number",
            "nullable": true
          },
          "mimeType": {
            "type": "string",
            "nullable": true
          },
          "parentId": {
            "type": "string",
            "nullable": true,
            "description": "The folder this result sits in"
          },
          "parentName": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "dataRoomId",
          "type",
          "name",
          "updatedAt",
          "parentId",
          "parentName"
        ]
      },
      "CreateFolderDto": {
        "type": "object",
        "properties": {
          "parentId": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "03 Legal"
          },
          "onConflict": {
            "type": "string",
            "enum": [
              "fail",
              "rename"
            ],
            "default": "fail",
            "description": "\"fail\" reports a conflict so the UI can ask; \"rename\" keeps both by appending a counter."
          }
        },
        "required": [
          "parentId",
          "name"
        ]
      },
      "RenameNodeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "onConflict": {
            "type": "string",
            "enum": [
              "fail",
              "rename"
            ],
            "default": "fail"
          }
        },
        "required": [
          "name"
        ]
      },
      "MoveNodeDto": {
        "type": "object",
        "properties": {
          "targetParentId": {
            "type": "string"
          },
          "onConflict": {
            "type": "string",
            "enum": [
              "fail",
              "rename"
            ],
            "default": "fail"
          }
        },
        "required": [
          "targetParentId"
        ]
      },
...
    }
  }
}